Se reflejan las calificaciones y horas en el mismo lugar.

+++ webMetics/Models/InscripcionModel.cs
@@ -20,6 +20,6 @@
 
         public int horasMatriculadas { get; set; }
 
-        public float? calificacion {  get; set; }
+        public double? calificacion {  get; set; }
     }
 }
